Fix malformed board edge fillet
These curves were not continuous, which created a crossover artifact where they met. This wasn't flagged as an issue because the board area is technically fully enclosed.

This commit fixes the curve so it's smooth. Note that my 1.2 boards manufactured by OSH Park turned out fine with the malformed edge, so this is mostly just a design-space issue.